Dependency Injection implemented
CLI processing is moved to another file.

New trait: IO. It is higher level than filesystem operations like write/read files. It is triggered by CLI as well as operations with tree triggered from methods like materialize, etc. IO trait works with images, folders, as well as meta and config.

New error type: TRIError. It describes all errors that could happen within TRI. Now all methods in IO as well as CLI and others return `Result<_, TRIError>`. No more .expect() in the middle of nowhere and unexpected panics. That will allow to test us easier.

Now there's type Logger. It's not a generic type and is passed around as it is. In the future there will be a field responsible for which levels (trace, info, warn, error) should be displayed.

All in all, it is now much better than it was before. Much more robust.