Adding flake (#10)
Now it's easier for developers and users to contribute to and use tri

In either way, first, install the `nix` package manager.

Run `nix develop` in the folder with the flake. Wait until it fetches all tooling, now you can develop stuff.

To run/test the app, run `nix shell` in the folder with the flake. Now the app is available as expected with the command `tri`, man page, autocompletion, whatever